web/HydrogenX/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2

# Hydrogen Chat 依赖（复用 chat/ 模块）
google-genai>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
tavily-python>=0.5.0
